requirements.txt
google-cloud-bigquery==2.34.4
Flask>= 2.0.0
gunicorn==20.1.0
pandas
pyarrow
pytest